Settings(gitsigns) : Update signs and use utils.map() function
Gitsigns recently set it's signs too big in my opinion.
This commit fixes this and use the same `text` option for each
sign.

Until now, I kept using a function for mappings that was only used
in gitsigns config (maybe this was taken from gitsigns doc).
As we already have our own map function in utils file, this was
unnecessary, so I removed gitsigns one to use the one from utils.